# openapi-specs/openapi.yaml
...
openapi: 3.0.3
components:
  schemas:
...
    DocumentResponse:
      type: object
      properties:
        id:
          format: int64
          type: integer
        title:
          type: string
        type:
          type: string
        summary:
          type: string
        characters:
          type: array
          items:
            type: string
        fullText:
          type: string
        yearPublished:
          format: int32
          type: integer
        authorName:
          type: string
        possibleQuestionsThisAnswers:
          type: array
          items:
            type: string
        createdAt:
          $ref: "#/components/schemas/OffsetDateTime"
        updatedAt:
          $ref: "#/components/schemas/OffsetDateTime"
...
    OffsetDateTime:
      format: date-time
      type: string
      example: 2022-03-10T12:15:50-04:00
    PagedResponseSectionResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SectionResponse"
        totalCount:
          format: int64
          type: integer
        currentPage:
          format: int32
          type: integer
        pageSize:
          format: int32
          type: integer
        totalPages:
          format: int32
          type: integer
        hasNextPage:
          type: boolean
        hasPreviousPage:
          type: boolean
    SectionResponse:
      type: object
      properties:
        id:
          format: int64
          type: integer
        chapterId:
          format: int64
          type: integer
        sequence:
          format: int32
          type: integer
        summary:
          type: string
        characters:
          type: array
          items:
            type: string
        fullText:
          type: string
        possibleQuestionsThisAnswers:
          type: array
          items:
            type: string
        createdAt:
          $ref: "#/components/schemas/OffsetDateTime"
        updatedAt:
          $ref: "#/components/schemas/OffsetDateTime"
...
    SemanticSearchMatchResponse:
      type: object
      properties:
        score:
          format: double
          type: number
        text:
          type: string
        sectionId:
          format: int64
          type: integer
        chapterId:
          format: int64
          type: integer
        documentId:
          format: int64
          type: integer
        bookTitle:
          type: string
        chapterLabel:
          type: string
...
paths:
...
  /rest/v1/documents/{id}:
    get:
      operationId: getDocumentById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      - name: include_full_text
        in: query
        schema:
          default: false
          type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentResponse"
      summary: Get Document By Id
      tags:
      - Document Resource
  /rest/v1/sections/by-chapter/{chapterId}:
    get:
      operationId: listSectionsByChapter
      parameters:
      - name: chapterId
        in: path
        required: true
        schema:
          format: int64
          type: integer
      - name: include_full_text
        in: query
        schema:
          default: false
          type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SectionResponse"
      summary: List By Chapter
      tags:
      - Section Resource
  /rest/v1/sections/by-document/{documentId}:
    get:
      operationId: listSectionsByDocument
      parameters:
      - name: documentId
        in: path
        required: true
        schema:
          format: int64
          type: integer
      - name: include_full_text
        in: query
        schema:
          default: false
          type: boolean
      - name: page
        in: query
        schema:
          format: int32
          default: 0
          type: integer
      - name: page_size
        in: query
        schema:
          format: int32
          default: 20
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResponseSectionResponse"
      summary: List By Document
      tags:
      - Section Resource
  /rest/v1/sections/by-sequence/{chapterId}/{sequenceNumber}:
    get:
      operationId: getSectionBySequence
      parameters:
      - name: chapterId
        in: path
        required: true
        schema:
          format: int64
          type: integer
      - name: sequenceNumber
        in: path
        required: true
        schema:
          format: int32
          type: integer
      - name: include_full_text
        in: query
        schema:
          default: false
          type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SectionResponse"
      summary: Get By Sequence
      tags:
      - Section Resource
  /rest/v1/sections/{id}:
    get:
      operationId: getSectionById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      - name: include_full_text
        in: query
        schema:
          default: false
          type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SectionResponse"
      summary: Get By Id
      tags:
      - Section Resource
  /rest/v1/semantic-search:
    get:
      operationId: searchSemantic
      parameters:
      - name: chapterId
        in: query
        schema:
          format: int64
          type: integer
      - name: documentId
        in: query
        schema:
          format: int64
          type: integer
      - name: maxResults
        in: query
        schema:
          format: int32
          default: 10
          type: integer
      - name: phrase
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SemanticSearchMatchResponse"
      summary: Search
      tags:
      - Semantic Search Resource
...
info:
  title: document-ingestion-api API
  version: 1.0.0-SNAPSHOT
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value